added data utils for interaction with torch's Dataloader
created a very basic torch Dataset that loads `hand1.json` multiple
times to build itself, as well as an example of a Dataloader function in
the tests that works with our model in a forward pass.

this dataloader uses a custom collate function to deal with
variable-lengthed inputs (because of action sequences).